changed int to size_t

+++ ft_buffer/ft_buffadd.c
@@ -6,7 +6,7 @@
 /*   By: yquaro <yquaro@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/02 12:36:28 by yquaro            #+#    #+#             */
-/*   Updated: 2019/11/16 16:03:20 by yquaro           ###   ########.fr       */
+/*   Updated: 2019/11/16 16:21:06 by yquaro           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,8 +22,8 @@
 
 void					ft_buffadd(t_buff *buff, const char *str)
 {
-	int					size_of_new_str;
-	int					remaining_size;
+	size_t				size_of_new_str;
+	size_t				remaining_size;
 
 	if (str == NULL)
 		return ;
@@ -50,8 +50,8 @@ void					ft_buffadd(t_buff *buff, const char *str)
 
 void					ft_buffladd(t_buff *buff, const char *str, int len)
 {
-	int					size_of_new_str;
-	int					remaining_size;
+	size_t				size_of_new_str;
+	size_t				remaining_size;
 
 	if (str == NULL)
 		return ;
